# typespec/tsp-output/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Call Booking API
  version: 0.0.0
tags: []
paths:
  /api/auth/login:
    post:
      operationId: Auth_login
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/auth/me:
    get:
      operationId: Auth_me
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      operationId: Auth_register
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/my/bookings:
    get:
      operationId: Bookings_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/BookingsListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: Bookings_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
  /api/my/bookings/{id}:
    delete:
      operationId: Bookings_cancel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /api/my/groups:
    get:
      operationId: Groups_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/GroupsListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: Groups_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisibilityGroup'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
  /api/my/groups/{id}:
    put:
      operationId: Groups_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/VisibilityGroup'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
    delete:
      operationId: Groups_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /api/my/groups/{id}/members:
    get:
      operationId: Groups_listMembers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/MembersListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: Groups_addMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
  /api/my/groups/{id}/members/{memberId}:
    delete:
      operationId: Groups_removeMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /api/my/schedules:
    get:
      operationId: Schedules_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SchedulesListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: Schedules_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
  /api/my/schedules/{id}:
    put:
      operationId: Schedules_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Schedule'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
    delete:
      operationId: Schedules_delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /api/users:
    get:
      operationId: Users_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/UsersListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}:
    get:
      operationId: Users_get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}/slots:
    get:
      operationId: Users_slots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SlotsListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
        userId:
          type: string
    AuthResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Booking:
      type: object
      required:
        - id
        - scheduleId
        - booker
        - owner
        - date
        - startTime
        - endTime
        - status
      properties:
        id:
          type: string
        scheduleId:
          type: string
        booker:
          $ref: '#/components/schemas/User'
        owner:
          $ref: '#/components/schemas/User'
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum:
            - active
            - cancelled
        createdAt:
          type: string
        cancelledAt:
          type: string
    BookingsListResponse:
      type: object
      required:
        - bookings
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
    CreateBookingRequest:
      type: object
      required:
        - ownerId
        - scheduleId
      properties:
        ownerId:
          type: string
        scheduleId:
          type: string
    CreateGroupRequest:
      type: object
      required:
        - name
        - visibilityLevel
      properties:
        name:
          type: string
        visibilityLevel:
          type: string
          enum:
            - family
            - work
            - friends
            - public
    CreateScheduleRequest:
      type: object
      required:
        - type
        - startTime
        - endTime
      properties:
        type:
          type: string
          enum:
            - recurring
            - one-time
        dayOfWeek:
          type: integer
          format: int32
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        isBlocked:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    GroupMember:
      type: object
      required:
        - id
        - groupId
        - member
        - addedBy
        - addedAt
      properties:
        id:
          type: string
        groupId:
          type: string
        member:
          $ref: '#/components/schemas/User'
        addedBy:
          type: string
        addedAt:
          type: string
    GroupsListResponse:
      type: object
      required:
        - groups
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/VisibilityGroup'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    MembersListResponse:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    Schedule:
      type: object
      required:
        - id
        - userId
        - type
        - startTime
        - endTime
        - isBlocked
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum:
            - recurring
            - one-time
        dayOfWeek:
          type: integer
          format: int32
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        isBlocked:
          type: boolean
        createdAt:
          type: string
    SchedulesListResponse:
      type: object
      required:
        - schedules
      properties:
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
    Slot:
      type: object
      required:
        - id
        - date
        - startTime
        - endTime
        - isBooked
      properties:
        id:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        isBooked:
          type: boolean
    SlotsListResponse:
      type: object
      required:
        - slots
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
    User:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    UsersListResponse:
      type: object
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    VisibilityGroup:
      type: object
      required:
        - id
        - ownerId
        - name
        - visibilityLevel
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        visibilityLevel:
          type: string
          enum:
            - family
            - work
            - friends
            - public
        createdAt:
          type: string
servers:
  - url: http://localhost:8080
    description: Local development server
    variables: {}